Use ContextVar colocated with model instead of adding top level export
