+++ src/agents/models/openai_chatcompletions.py
@@ -24,9 +24,8 @@ from ..tracing.span_data import GenerationSpanData
 from ..tracing.spans import Span
 from ..usage import Usage
 from ..util._json import _to_dump_compatible
-from . import _openai_shared
 from .chatcmpl_converter import Converter
-from .chatcmpl_helpers import HEADERS, ChatCmplHelpers
+from .chatcmpl_helpers import HEADERS, USER_AGENT_OVERRIDE, ChatCmplHelpers
 from .chatcmpl_stream_handler import ChatCmplStreamHandler
 from .fake_id import FAKE_RESPONSES_ID
 from .interface import Model, ModelTracing
@@ -353,6 +352,7 @@ class OpenAIChatCompletionsModel(Model):
 
     def _merge_headers(self, model_settings: ModelSettings):
         merged = {**HEADERS, **(model_settings.extra_headers or {})}
-        if ua_override := _openai_shared.get_user_agent_override():
-            merged["User-Agent"] = ua_override
+        ua_ctx = USER_AGENT_OVERRIDE.get()
+        if ua_ctx is not None:
+            merged["User-Agent"] = ua_ctx
         return merged
